+++ test/Utils/EnumExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.ComponentModel;
+using System.Reflection;
 
 namespace FluentlyHttpClient.Test.Utils
 {
@@ -9,11 +10,10 @@ namespace FluentlyHttpClient.Test.Utils
 		{
 			var fi = value.GetType().GetField(value.ToString());
 
-			var attributes = (DescriptionAttribute[])fi.GetCustomAttributes(typeof(DescriptionAttribute), false);
+			var attributes = fi.GetCustomAttribute<DescriptionAttribute>(false);
 
-			if (attributes != null &&
-				attributes.Length > 0)
-				return attributes[0].Description;
+			if (attributes != null)
+				return attributes.Description;
 			return value.ToString();
 		}
 	}